Mark version 3.3.3's tracking of anonymous statistics as a breaking change
After upgrading from 3.2.x to 3.3.x and scanning the changelog for breaking changes, this change broke automated deployment via our continuous deployment. I had to run a manual deploy to create the settings file, after which our CI server could deploy again.

I'm baffled that this was not marked as a breaking change, but I hope to rectify it after the fact with this commit.