Return longer SHA1 hash when fetching current revision
The mechanism Capistrano used to get the current commit hash of a
repository only returned the first 7 characters of the hash (thanks to the
`--abbrev-commit` flag).

While usually more convenient, the problem with a 7 character
abbreviation is that by the time a repository reaches 10,000 commits, it
becomes quite likely that you'll have a collision in the first 7
characters of two commits (~17% chance).

A collision is problematic if you intend to use these hashes to generate
links in emails or chat messages notifying your team of a deploy.

Increasing the minimum number of characters to 12 keeps the hash short
but virtually eliminates the possibility of a collision for all
practical purposes.